+++ CHANGELOG.md
@@ -13,7 +13,9 @@
   * [`Yggdrasil`] Now `subscribe/1`, `unsubscribe/1` and `publish/3` can
   receive a map or a `Keyword` list as channel as well as the
   `Yggdrasil.Channel` struct.
-  * Added Yggdrasil logo.
+  * Added Yggdrasil logo to documentation.
+  * Added Yggdrasil Ethereum adapter to documentation.
+  * Added Yggdrasil GraphQL adapter to documentation.
 
 # Changelog for 4.1.0